+++ src/models/scheduler.ts
@@ -241,7 +241,7 @@ export type SchedulerBookingProperties = {
   calendarInviteToGuests?: boolean;
   cancellationPolicy?: string;
   confirmationEmailsToGuests?: boolean;
-  confirmationEmailToHost?: boolean;
+  confirmationEmailsToHost?: boolean;
   confirmationMethod?: string;
   minBookingNotice?: number;
   minBuffer?: number;
@@ -260,7 +260,7 @@ export class SchedulerBooking extends Model
   calendarInviteToGuests?: boolean;
   cancellationPolicy?: string;
   confirmationEmailsToGuests?: boolean;
-  confirmationEmailToHost?: boolean;
+  confirmationEmailsToHost?: boolean;
   confirmationMethod?: string;
   minBookingNotice?: number;
   minBuffer?: number;
@@ -295,8 +295,8 @@ export class SchedulerBooking extends Model
       modelKey: 'confirmationEmailsToGuests',
       jsonKey: 'confirmation_emails_to_guests',
     }),
-    confirmationEmailToHost: Attributes.Boolean({
-      modelKey: 'confirmationEmailToHost',
+    confirmationEmailsToHost: Attributes.Boolean({
+      modelKey: 'confirmationEmailsToHost',
       jsonKey: 'confirmation_emails_to_host',
     }),
     confirmationMethod: Attributes.String({