+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 7.0.0 / TBD
 * **BREAKING CHANGE**: Convert `Nylas` class from a static to a non-static class
+* **BREAKING CHANGE**: Rename `confirmationEmailToHost` to `confirmationEmailsToHost` in `SchedulerBooking`
 
 ### 6.8.0 / 2023-02-03
 * Local webhook testing support